requirements_free.txt
# 100% Free Market Data Solution - No Paid APIs Required
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
schedule==1.2.0

# Optional for enhanced functionality (all free)
supabase==1.0.4
flask==2.3.3
flask-cors==4.0.0